+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+
+### Added
+
+- Add `set_remote_url` git method and branch as the input parameter of `list_commits` ([84])
+
+### Changed
+
+### Fixed
+
+- Fix branch exists git function ([82])
+
+[84]: https://github.com/openlawlibrary/taf/pull/84
+[82]: https://github.com/openlawlibrary/taf/pull/82
+
 ## [0.2.0] - 11/30/2019
 
 ### Added
@@ -18,11 +32,9 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Fixed
 
-- Fix branch exists git function ([82])
 - Creation of new repositories made more robust ([79])
 
 
-[82]: https://github.com/openlawlibrary/taf/pull/82
 [79]: https://github.com/openlawlibrary/taf/pull/79