+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+## [0.2.1] - 12/19/2019
 
 ### Added
 
@@ -109,7 +110,8 @@ and this project adheres to [Semantic Versioning][semver].
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/
 [semver]: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/openlawlibrary/pygls/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/openlawlibrary/pygls/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/openlawlibrary/taf/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/openlawlibrary/taf/compare/v0.1.8...v0.2.0
 [0.1.8]: https://github.com/openlawlibrary/taf/compare/v0.1.7...v0.1.8
 [0.1.7]: https://github.com/openlawlibrary/taf/compare/v0.1.6...v0.1.7